Move viewport code to its own source files.
